requirements.txt
pyserial==3.4
argparse==1.4.0